requirements.txt
alembic==1.18.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
Django==6.0.1
django-filter==25.2
django-phonenumber-field==8.4.0
django-resized==1.0.3
djangorestframework==3.16.1
fastapi==0.128.0
greenlet==3.3.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
phonenumbers==9.0.22
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
